[dg] verb-first env cli commands
