+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/test_list_commands.py
@@ -9,6 +9,8 @@ from dagster_dg.utils import ensure_dagster_dg_tests_import
 
 ensure_dagster_dg_tests_import()
 
+from dagster_dg.utils import ensure_dagster_dg_tests_import
+
 from dagster_dg_tests.utils import (
     ProxyRunner,
     assert_runner_result,
@@ -432,3 +434,37 @@ def _sample_env_var_assets():
     @asset(kinds={"sling"}, group_name=os.getenv("GROUP_NAME", "MISSING"))
     def alpha():
         pass
+
+
+# ########################
+# ##### LIST
+# ########################
+
+
+def test_list_env_succeeds():
+    with (
+        ProxyRunner.test(use_fixed_test_components=True) as runner,
+        isolated_example_project_foo_bar(runner, in_workspace=False),
+    ):
+        result = runner.invoke("list", "env")
+        assert_runner_result(result)
+        assert (
+            result.output.strip()
+            == textwrap.dedent("""
+            No environment variables are defined for this project.
+        """).strip()
+        )
+
+        Path(".env").write_text("FOO=bar")
+        result = runner.invoke("list", "env")
+        assert_runner_result(result)
+        assert (
+            result.output.strip()
+            == textwrap.dedent("""
+               ┏━━━━━━━━━┳━━━━━━━┓
+               ┃ Env Var ┃ Value ┃
+               ┡━━━━━━━━━╇━━━━━━━┩
+               │ FOO     │ bar   │
+               └─────────┴───────┘
+        """).strip()
+        )
